Added a lot of older NeXus material. Most of it is done with latex beamer. pdflatex is your freind when building it.
The .dia files are the sources for artwork generated with the Dia diagram editor on linux
Added a lot of older NeXus material. Most of it is done with latex beamer.
pdflatex is your freind when building it.

The .dia files are the sources for artwork generated with the Dia diagram editor on linux