Changes to Section IIIA (mostly)
The most substantive changes are
(i) shift NXdata description above NXmonitor, since it’s more
important.
(ii) make it clear that not all data in NXdata groups are linked.
(iii) describe the NXmonitor group, without apologizing for its
location in the NXentry groups (since I don’t feel an apology is
necessary, and it’s probably unwise when you are promoting the format).
This is the right place for, e.g., all pulsed neutron experiments.